Изменил коды ответов на корректные. Теперь у разных методов на 1 урл свои собственные handler-ы. Добавил обработку ошибок ко всем возможным функциям. Изменил способ приведения к json-у в коде. Разнес по разным файлам роутинг, модели и main сервер. Добавил чтение секрета из пременных окружения.
